fix(admin): move MainPageLocation validation to form to unblock M2M updates
